pyproject.toml
[project]
name = "simple-agent-memory"
version = "0.1.0"
description = "Plug-and-play memory system for AI agents"
requires-python = ">=3.11"
dependencies = ["numpy", "aiosqlite"]

[project.optional-dependencies]
openai = ["openai"]
anthropic = ["anthropic"]
gemini = ["google-generativeai"]
http = ["httpx"]
all = ["openai", "anthropic", "google-generativeai", "httpx"]
dev = ["pytest", "pytest-asyncio"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
